moved the code into a separate namespace/package
