--- src/main/java/org/got5/tapestry5/jquery/AjaxUploadConstants.java
@@ -1,9 +1,0 @@
-package org.got5.tapestry5.jquery;
-
-public class AjaxUploadConstants {
-
-    /**
-     * Parameter used by the fileuploader js
-     */
-    public static final String FILE_UPLOAD_PARAMETER = "qqfile";
-}
