--- src/main/java/org/got5/tapestry5/jquery/services/parsers/RequestParserSource.java
@@ -1,6 +1,0 @@
-package org.got5.tapestry5.jquery.services.parsers;
-
-public interface RequestParserSource {
-
-    RequestParser createRequestParser();
-}
