+++ src/main/java/org/got5/tapestry5/upload/AjaxUploadSymbolConstants.java
@@ -1,0 +1,27 @@
+// Copyright 2007, 2008 The Apache Software Foundation
+//
+// Licensed under the Apache License, Version 2.0 (the "License");
+// you may not use this file except in compliance with the License.
+// You may obtain a copy of the License at
+//
+//     http://www.apache.org/licenses/LICENSE-2.0
+//
+// Unless required by applicable law or agreed to in writing, software
+// distributed under the License is distributed on an "AS IS" BASIS,
+// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+// See the License for the specific language governing permissions and
+// limitations under the License.
+
+package org.got5.tapestry5.upload;
+
+/**
+ * Configuration symbols.
+ */
+public final class AjaxUploadSymbolConstants {
+
+    public static final String ASSETS_ROOT = "tjq-upload.assets-root";
+
+    public static final String ASSETS_ROOT_VALUE = "${" + ASSETS_ROOT + "}";
+
+    private AjaxUploadSymbolConstants() { }
+}
