--- src/test/java/org/got5/tapestry5/jquery/pages/Index.java
@@ -1,8 +1,0 @@
-package org.got5.tapestry5.jquery.pages;
-
-
-
-public class Index {
-	
-
-}
