+++ src/test/java/org/got5/tapestry5/upload/pages/AjaxUploadTest.java
@@ -1,4 +1,4 @@
-package org.got5.tapestry5.jquery.pages;
+package org.got5.tapestry5.upload.pages;
 
 import java.io.File;
 import java.util.ArrayList;
@@ -16,7 +16,7 @@ import org.apache.tapestry5.ioc.Messages;
 import org.apache.tapestry5.ioc.annotations.Inject;
 import org.apache.tapestry5.json.JSONObject;
 import org.apache.tapestry5.services.ajax.AjaxResponseRenderer;
-import org.got5.tapestry5.jquery.AjaxUploadEventConstants;
+import org.got5.tapestry5.upload.AjaxUploadEventConstants;
 
 public class AjaxUploadTest {
