+++ src/test/java/org/got5/tapestry5/upload/pages/Index.java
@@ -1,0 +1,8 @@
+package org.got5.tapestry5.upload.pages;
+
+
+
+public class Index {
+	
+
+}
