+++ src/test/resources/org/got5/tapestry5/upload/pages/AjaxUploadTest_de.properties
